requirements.txt
markdown
Pygments
python-slugify
